Updated countries endpoint to use new query, fixed locations bug
The temp way to determine is monitor was returning nulls when there is
no origin value
